requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
instructor==1.0.0
openai==1.12.0
marker-pdf==0.3.1
torch==2.4.1
Pillow==10.2.0
